+++ src/config/locales/en-US.json
@@ -745,6 +745,7 @@
   },
   "qr":{
     "qr_scan": "QR Scan",
-    "open": "Open URL"
+    "open": "Open URL",
+    "detected_url": "Detected URL"
   }
 }
